Rename `more-dropdown` to `more-dropdown-links`
