+++ source/features/remove-unused-repo-tabs.tsx
@@ -9,8 +9,8 @@ import getTabCount from '../github-helpers/get-tab-count';
 import looseParseInt from '../helpers/loose-parse-int';
 import {getWorkflows} from './next-scheduled-github-action';
 import abbreviateNumber from '../helpers/abbreviate-number';
-import {onlyShowInDropdown} from './more-dropdown';
 import {buildRepoURL, getRepo} from '../github-helpers';
+import {onlyShowInDropdown, unhideOverflowDropdown} from './more-dropdown-links';
 
 function mustKeepTab(tab: HTMLElement | undefined): boolean {
 	return (
@@ -83,7 +83,20 @@ async function initProjects(): Promise<void | false> {
 	projectsTab!.remove();
 }
 
+async function init(): Promise<void> {
+	// The user may have disabled `more-dropdown` so un-hide it
+	await unhideOverflowDropdown();
+	onlyShowInDropdown('security-tab');
+	onlyShowInDropdown('insights-tab');
+}
+
 void features.add(__filebasename, {
+	include: [
+		pageDetect.isRepo,
+	],
+	awaitDomReady: false,
+	init,
+}, {
 	include: [
 		pageDetect.isRepo,
 		pageDetect.isOrganizationProfile,
