+++ source/refined-github.ts
@@ -37,7 +37,7 @@ import './features/remove-diff-signs';
 
 import './features/useful-not-found-page';
 import './features/trending-menu-item';
-import './features/more-dropdown';
+import './features/more-dropdown-links';
 import './features/releases-tab';
 import './features/comment-fields-keyboard-shortcuts';
 import './features/one-key-formatting';
